requirements.txt
dvc[gs]
protobuf==4.24.4
Flask
pandas~=2.1.3
numpy~=1.26.1
scikit-learn~=1.3.2
scipy
openpyxl~=3.1.2
gcsfs~=2023.10.0
pytz~=2023.3.post1
joblib~=1.3.2
python-dotenv==1.0.0
apache-airflow
google-cloud-aiplatform==1.36.3
grpcio==1.51.3
airflow